# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RuiQi 应用防火墙管理系统 API",
        "title": "RuiQi-WAF API",
        "termsOfService": "https://github.com/HUAHUAI23/RuiQi",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/HUAHUAI23/RuiQi",
            "email": "huahua1319873800@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:2333",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/stats/combined-time-series": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "同时获取请求数和拦截数的时间序列数据，用于图表展示",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计信息"
                ],
                "summary": "获取组合时间序列数据",
                "parameters": [
                    {
                        "enum": [
                            "24h",
                            "7d",
                            "30d"
                        ],
                        "type": "string",
                        "default": "24h",
                        "description": "时间范围：24h(24小时)、7d(7天)、30d(30天)",
                        "name": "timeRange",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取组合时间序列数据成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CombinedTimeSeriesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/stats/overview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定时间范围内的统计概览数据，包括请求数、流量、错误率等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计信息"
                ],
                "summary": "获取统计概览数据",
                "parameters": [
                    {
                        "enum": [
                            "24h",
                            "7d",
                            "30d"
                        ],
                        "type": "string",
                        "default": "24h",
                        "description": "时间范围：24h(24小时)、7d(7天)、30d(30天)",
                        "name": "timeRange",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取统计概览成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OverviewStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/stats/realtime-qps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取最近的实时QPS数据点",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计信息"
                ],
                "summary": "获取实时QPS数据",
                "parameters": [
                    {
                        "maximum": 60,
                        "minimum": 1,
                        "type": "integer",
                        "default": 30,
                        "description": "返回的数据点数量，默认30个点，最大60个点",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取实时QPS数据成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RealtimeQPSResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/stats/time-series": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定时间范围和指标类型的时间序列数据，用于图表展示",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计信息"
                ],
                "summary": "获取时间序列数据",
                "parameters": [
                    {
                        "enum": [
                            "24h",
                            "7d",
                            "30d"
                        ],
                        "type": "string",
                        "default": "24h",
                        "description": "时间范围：24h(24小时)、7d(7天)、30d(30天)",
                        "name": "timeRange",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "requests",
                            "blocks"
                        ],
                        "type": "string",
                        "default": "requests",
                        "description": "指标类型：requests(请求数)、blocks(拦截数)",
                        "name": "metric",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取时间序列数据成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TimeSeriesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/stats/traffic-time-series": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定时间范围的入站和出站流量时间序列数据，用于图表展示",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计信息"
                ],
                "summary": "获取流量时间序列数据",
                "parameters": [
                    {
                        "enum": [
                            "24h",
                            "7d",
                            "30d"
                        ],
                        "type": "string",
                        "default": "24h",
                        "description": "时间范围：24h(24小时)、7d(7天)、30d(30天)",
                        "name": "timeRange",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取流量时间序列数据成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TrafficTimeSeriesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
...
        "/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取HAProxy原始的统计信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计信息"
                ],
                "summary": "获取HAProxy原始统计数据",
                "responses": {
                    "200": {
                        "description": "获取统计数据成功",
                        "schema": {
                            "$ref": "#/definitions/model.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CombinedTimeSeriesResponse": {
            "description": "同时包含请求数和拦截数的时间序列数据",
            "type": "object",
            "properties": {
                "blocks": {
                    "description": "拦截数时间序列",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.TimeSeriesResponse"
                        }
                    ]
                },
                "requests": {
                    "description": "请求数时间序列",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.TimeSeriesResponse"
                        }
                    ]
                },
                "timeRange": {
                    "description": "时间范围",
                    "type": "string",
                    "example": "24h"
                }
            }
        },
...
        "dto.OverviewStats": {
            "description": "概览统计数据，包含各项关键指标",
            "type": "object",
            "properties": {
                "attackIPCount": {
                    "description": "攻击IP数量",
                    "type": "integer",
                    "example": 45
                },
                "blockCount": {
                    "description": "安全统计",
                    "type": "integer",
                    "example": 123
                },
                "error4xx": {
                    "description": "错误统计",
                    "type": "integer",
                    "example": 234
                },
                "error4xxRate": {
                    "description": "4xx错误率(百分比)",
                    "type": "number",
                    "example": 0.23
                },
                "error5xx": {
                    "description": "5xx错误数量",
                    "type": "integer",
                    "example": 45
                },
                "error5xxRate": {
                    "description": "5xx错误率(百分比)",
                    "type": "number",
                    "example": 0.05
                },
                "inboundTraffic": {
                    "description": "入站流量(字节)",
                    "type": "integer",
                    "example": 67890123
                },
                "maxQPS": {
                    "description": "最大QPS",
                    "type": "integer",
                    "example": 150
                },
                "outboundTraffic": {
                    "description": "出站流量(字节)",
                    "type": "integer",
                    "example": 12345678
                },
                "timeRange": {
                    "description": "时间范围",
                    "type": "string",
                    "example": "24h"
                },
                "totalRequests": {
                    "description": "流量统计",
                    "type": "integer",
                    "example": 123456
                }
            }
        },
        "dto.RealtimeQPSData": {
            "description": "实时QPS数据点，包含时间戳和值",
            "type": "object",
            "properties": {
                "timestamp": {
                    "description": "时间戳",
                    "type": "string",
                    "example": "2024-01-01T12:30:45Z"
                },
                "value": {
                    "description": "QPS值",
                    "type": "integer",
                    "example": 120
                }
            }
        },
        "dto.RealtimeQPSResponse": {
            "description": "实时QPS数据响应",
            "type": "object",
            "properties": {
                "data": {
                    "description": "QPS数据点列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RealtimeQPSData"
                    }
                }
            }
        },
...
        "dto.TimeSeriesDataPoint": {
            "description": "时间序列图表数据点",
            "type": "object",
            "properties": {
                "timestamp": {
                    "description": "时间戳，表示数据点的时间",
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "value": {
                    "description": "数值，表示该时间点的指标值",
                    "type": "integer",
                    "example": 128
                }
            }
        },
        "dto.TimeSeriesResponse": {
            "description": "时间序列图表数据响应",
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据点列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TimeSeriesDataPoint"
                    }
                },
                "metric": {
                    "description": "指标名称",
                    "type": "string",
                    "example": "requests"
                },
                "timeRange": {
                    "description": "时间范围",
                    "type": "string",
                    "example": "24h"
                }
            }
        },
        "dto.TrafficDataPoint": {
            "description": "流量时间序列图表数据点",
            "type": "object",
            "properties": {
                "inboundTraffic": {
                    "description": "入站流量(字节)",
                    "type": "integer",
                    "example": 1024000
                },
                "outboundTraffic": {
                    "description": "出站流量(字节)",
                    "type": "integer",
                    "example": 2048000
                },
                "timestamp": {
                    "description": "时间戳，表示数据点的时间",
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                }
            }
        },
        "dto.TrafficTimeSeriesResponse": {
            "description": "流量时间序列图表数据响应",
            "type": "object",
            "properties": {
                "data": {
                    "description": "流量数据点列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TrafficDataPoint"
                    }
                },
                "timeRange": {
                    "description": "时间范围",
                    "type": "string",
                    "example": "24h"
                }
            }
        },
...
        "model.ErrResponse": {
            "description": "错误的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "参数错误"
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.ErrResponseDontShowError": {
            "description": "错误的API响应标准格式,不展示 error",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
        "model.SuccessResponse": {
            "description": "成功的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer {token} 格式进行身份验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}